他人のitemを削除不可に設定

+++ src/items/items.service.ts
@@ -9,6 +9,7 @@ import { ItemStatus } from './item-status.enum'
 import { Item } from '../entities/item.entity'
 import { ItemRepository } from './item.repository'
 import { User } from 'src/entities/user.entity'
+import { GetUser } from 'src/auth/decorator/get-user.decorator'
 
 @Injectable()
 export class ItemsService {
@@ -41,7 +42,11 @@ export class ItemsService {
     return await this.itemRepository.save(item)
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.findById(id)
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません')
+    }
     await this.itemRepository.delete(id)
   }
 }
